feat: moved parseFunction into the individual ir file
