+++ src/ir/ir.h
@@ -29,10 +29,4 @@ void pushInstruction(IR_BASIC_BLOCK* block, IR_INSTRUCTION* instruction);
  */
 void appendInstruction(IR_BASIC_BLOCK* block, IR_INSTRUCTION_CODE code, unsigned char params[], int paramsCount);
 
-/**
- * Parses a AST function into IR.
- * @param node the AST node representing the function.
- */
-IR_FUNCTION parseFunction(AST_FUNCTION_DEC* node);
-
 #endif
